requirements.txt
aiopg==1.3.4
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
autobahn==22.7.1
Automat==20.2.0
black==22.8.0
CacheControl==0.13.1
cachetools==5.3.1
certifi==2022.6.15
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==2.1.1
click==8.1.3
colander==1.8.3
colorama==0.4.5
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
daphne==3.0.2
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.6
Django==3.2.15
django-admin-honeypot==1.1.0
django-allauth==0.51.0
django-cors-headers==3.13.0
django-filter==22.1
django-pytest==0.2.0
django-rest-auth==0.9.5
django-taggit==3.0.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
drf-yasg==1.21.5
exceptiongroup==1.1.1
factory-boy==3.2.1
Faker==18.4.0
fcm-django==2.0.0
filelock==3.12.0
firebase-admin==6.1.0
geographiclib==1.52
geopy==2.2.0
google-api-core==2.11.0
google-api-python-client==2.88.0
google-auth==2.19.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
grpcio==1.54.2
grpcio-status==1.54.2
gunicorn==20.1.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==6.4.1
incremental==21.3.0
inflection==0.5.1
iniconfig==1.1.1
iso8601==1.1.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
msgpack==1.0.4
mypy-extensions==0.4.3
oauthlib==3.2.0
openapi-codec==1.3.2
packaging==23.1
pathspec==0.10.1
Pillow==9.2.0
pipenv==2023.4.20
platformdirs==3.2.0
pluggy==1.0.0
proto-plus==1.22.2
protobuf==4.23.2
psycopg2-binary==2.9.3
py==1.11.0
Py-Authorize==1.4.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.7.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
pytest==7.3.1
pytest-django==4.5.2
pytest-factoryboy==2.5.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.2.1
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
service-identity==21.1.0
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
translationstring==1.4
twilio==7.14.1
Twisted==22.8.0
txaio==22.2.1
typed-ast==1.5.4
typing_extensions>=3.6.5,<4.6
uritemplate==4.1.1
urllib3==1.26.12
virtualenv==20.22.0
virtualenv-clone==0.5.7
vistir==0.8.0
wrapt==1.14.1
zipp==3.8.1
zope.interface==5.4.0
